a few more cleanups, and a fix for Bitstream::SendString
git-svn-id: svn://svn.code.sf.net/p/segs/code2@279 a19be42f-918d-4297-8077-ab25d1e4f51f

+++ Projects/CoX/Servers/MapServer/src/Events/InputState.cpp
@@ -139,7 +139,7 @@ void InputState::recv_client_opts(BitStream &bs)
     entry=opts.get(opt_idx);
     while(entry!=0)
     {
-        for(int i=0; i<entry->m_args.size(); i++)
+        for(size_t i=0; i<entry->m_args.size(); i++)
         {
             ClientOption::Arg &arg=entry->m_args[i];
             switch ( arg.type )
